requirements.txt
aiodogstatsd==0.16.0.post0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
argcomplete==3.5.3
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
billiard==4.2.1
boto3==1.36.2
botocore==1.36.2
celery==5.4.0
certifi==2024.12.14
cfgv==3.4.0
cfn-flip==1.3.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
defusedxml==0.7.1
dill==0.3.9
distlib==0.3.9
Django==5.1.5
django-crispy-forms==2.3
django-filter==24.3
django-ses==4.3.2
django-sslserver==0.22
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
durationpy==0.9
exceptiongroup==1.2.2
filelock==3.17.0
flower==2.0.1
frozenlist==1.5.0
hjson==3.1.0
httpie==3.2.4
humanize==4.11.0
identify==2.6.6
idna==3.10
imageio==2.37.0
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kappa==0.6.0
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
mdurl==0.1.2
multidict==6.1.0
multiprocess==0.70.17
mysqlclient==2.2.7
nodeenv==1.9.1
numpy==2.2.2
packaging==24.2
pillow==11.1.0
placebo==0.9.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
Pygments==2.19.1
PyJWT==2.10.1
pyotp==2.9.0
PySocks==1.7.1
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-magic-bin==0.4.14
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.11.1
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
troposphere==4.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
virtualenv==20.29.1
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.18.3
